make browser tests work again
