+++ hyperspace.browser.js
@@ -1,0 +1,3 @@
+module.exports = async function createBackend () {
+  throw new Error('Hyperspace backend is not supported in browsers')
+}
