+++ test/browser.js
@@ -1,0 +1,4 @@
+const run = require('./tests')
+const createNative = require('./lib/native')
+
+run(createNative, 'native')
